% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average.R
\name{average}
\alias{average}
\title{Average}
\usage{
average(vector)
}
\arguments{
\item{vector}{A numeric vector}
}
\value{
A numeric value
}
\description{
Does some complex math to get an average value
}
\examples{
average(rnorm(10))

}
\author{
Thomas Guillerme
}
