% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moms-package.R
\name{Demo_data}
\alias{Demo_data}
\title{Demo datasets}
\format{@@}
\source{
@@
}
\description{
A set six trait spaces with different groups and different dimensions
}
\details{
\itemize{
  \item \code{Beck2014} @@
  \item \code{Wright2017} @@
  \item \code{Marcy2016} @@
  \item \code{@@} @@
  \item \code{Jones2015} @@
  \item \code{@@} @@
}
}
\references{
@@
}
