% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdp.R
\name{dispRity.vdp}
\alias{dispRity.vdp}
\title{disparity vdp}
\usage{
dispRity.vdp(vdp, volume, density, position, base.relative = TRUE)
}
\arguments{
\item{vdp}{A list output from \code{\link{make.vdp}}}

\item{volume}{a volume function as passed to \code{\link[dispRity]{dispRity}}}

\item{density}{a density function as passed to \code{\link[dispRity]{dispRity}}}

\item{position}{a position function as passed to \code{\link[dispRity]{dispRity}}}

\item{base.relative}{whether to make the results relative to the base (\code{TRUE} - default) or absolute (\code{FALSE})}
}
\description{
Measure disparity on a vdp object
}
\examples{
## Make a Volume/density/position list
vdp_list <- make.vdp()

## Plotting the transformations
plot.vdp(vdp_list)

## Calculate disparity
vdp_disp <- dispRity.vdp(vdp_list, volume = c(prod, ranges),
                                   density = c(mean, neighbours),
                                   position = c(mean, displacements))

## Plotting the results with disparity
plot.vdp(vdp_list, disparity = vdp_disp)

}
\seealso{
\code{\link{make.vdp}}, \code{\link{plot.vdp}}
}
\author{
Thomas Guillerme
}
