% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.results.R
\name{extract.results}
\alias{extract.results}
\title{Extract results}
\usage{
extract.results(all_results, metric, element = "all", transfo = "all")
}
\arguments{
\item{all_results}{the output from a simulation bundle}

\item{metric}{which metric to extract}

\item{element}{which element to extract (e.g. number of dimensions, type of space, etc.) can be "all"}

\item{transfo}{which transformation to extract (e.g. random, density, etc.) can be "all"}
}
\description{
Extract specific results from the simulation lists
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
