% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.metrics.R
\name{fast.disparity}
\alias{fast.disparity}
\title{Fast disparity}
\usage{
fast.disparity(group, space, metric, rare.dim)
}
\arguments{
\item{group}{The elements to select in space}

\item{space}{The space}

\item{metric}{The metric (in dispRity format)}

\item{rare.dim}{The dimensions to select in space}
}
\description{
Fast disparity calculation (skips all the sanitizing and formatting of dispRity)
}
\details{
Prefer using the proper \code{\link[dispRity]{dispRity}} function.
This function is a quick shortcut that can create unwanted problems in disparity pipelines.
}
