% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOconts-functions.R
\name{findDes}
\alias{findDes}
\title{Find Multi-Outcome Multi-Stage Trials That Allow a General Number of Efficacious Outcomes}
\usage{
findDes(
  K = default.K,
  m = default.m,
  J = default.J,
  corr.mat = NULL,
  vars = NULL,
  corr.scalar = NULL,
  nsims = 1e+05,
  wang.delta = 0,
  alpha = default.alpha,
  power = default.power,
  delta0 = default.delta0,
  delta1 = default.delta1,
  delta.true = NULL,
  reuse.true.deltas = TRUE,
  vars.true = NULL,
  working.outs = NULL,
  maxn.stage = 200,
  return.boundaries = FALSE,
  return.ts = FALSE
)
}
\description{
This function allows users to find a single-arm multi-outcome multi-stage trial that
allows ending for trial success if promising effects are observed on a general number
of outcomes, specified by the user.
}
