% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOconts-functions.R
\name{findWangTsiatisBounds}
\alias{findWangTsiatisBounds}
\title{Finds Wang & Tsiatis Stopping Boundaries}
\usage{
findWangTsiatisBounds(C, delta, J)
}
\arguments{
\item{C}{Constant}

\item{delta}{Parameter determining shape of stopping regions.}

\item{J}{Number of stages}
}
\value{
Returns a data frame of J rows and two columns, representing the lower and upper stopping boundaries for each of the J stages.
}
\description{
This function finds stopping boundaries using the formula of Wang & Tsiatis
}
