% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.fable.R
\name{generate.fable.empirical}
\alias{generate.fable.empirical}
\title{Generates fable plots for the empirical data}
\usage{
generate.fable.empirical(data, test, plot.param, precision = 1e-16)
}
\arguments{
\item{data}{the disparity data}

\item{test}{the associated test with the data}

\item{plot.param}{the plot parameters}

\item{precision}{the value under which to consider values as zeros}
}
\description{
Generates the plots for the empirical example fable. The R markdown compiler should name them as "<snippet_name>-<ID>.<type>"
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
