% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.fable.R
\name{generate.fable.plot}
\alias{generate.fable.plot}
\title{Generates fable plots}
\usage{
generate.fable.plot(data, metric, what)
}
\arguments{
\item{data}{the disparity data}

\item{metric}{the metric ID}

\item{what}{the name of the chain as given pas the code snippet generating the plots}

\item{resolution}{the plot resolution in the fable}
}
\description{
Generates the plots for the fable. The R markdown compiler should name them as "<snippet_name>-<ID>.<type>"
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
