% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.simulation.R
\name{metrics.simulation}
\alias{metrics.simulation}
\title{Apply metrics}
\usage{
metrics.simulation(space, reduction, metrics, rare.dim = NULL, verbose = TRUE)
}
\arguments{
\item{space}{a list of spaces}

\item{reduction}{a list of reductions}

\item{metrics}{a list of metrics}

\item{rare.dim}{a rarefaction value (optional)}

\item{verbose}{whether to be verbose or not}
}
\description{
Apply the metrics to the simulated spaces
}
\examples{
## Generate a space
space <- space.simulation(distributions = list(distribution = rnorm),
                         dimensions = 2, elements = 20, replicates = 3)

## Make the shifts
groups <- lapply(space, shift.group.simulation, remove = 0.5)

## List of metrics
metrics <- list(c(sum, variances), c(sum, ranges))

## Apply the metrics
metrics.simulation(space, groups, metrics)

}
\seealso{
space.simulation shift.group.simulation
}
\author{
Thomas Guillerme
}
