% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moms-package.R
\docType{package}
\name{moms-package}
\alias{moms-package}
\title{Measuring Occupancy of Multidimensional Spaces}
\description{
A shiny app for measuring occupancy of multidimensional spaces
}
\author{
Thomas Guillerme <guillert@tcd.ie>
}
