% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.fable.R
\name{plot.id}
\alias{plot.id}
\title{Plotting distribution by ID}
\usage{
\method{plot}{id}(plot.ID, path = "fig.path", resolution = 600)
}
\arguments{
\item{plot.ID}{the number of the plot as generated by generate the R markdown compiler}

\item{path}{where the plots are generated by the compiler}

\item{resolution}{the plot resolution in the fable}

\item{chain}{the name of the chain as given pas the code snippet generating the plots}

\item{type}{the file type (e.g. ".png")}
}
\description{
Plots one distribution for the fable following the plot ID
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
