% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.utilities.R
\name{plot.metrics}
\alias{plot.metrics}
\title{Plot the results fable for the metrics analysis}
\usage{
\method{plot}{metrics}(cent.tend, CIs, col, space, remove_list, defaults)
}
\arguments{
\item{cent.tend}{the analysis central tendencies}

\item{CIs}{the analysis CIs}

\item{col}{the colours for the results (i.e. for the different metrics)}

\item{space}{the space to be plotted on the side (plot.space)}

\item{remove_list}{a list of elements to remove in the space (plot.space)}

\item{defaults}{a list of arguments for plot.space}
}
\description{
plotting the spaces changing
}
