% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.utilities.R
\name{plot.space}
\alias{plot.space}
\title{Plot space}
\usage{
\method{plot}{space}(space, remove, main, defaults, ...)
}
\arguments{
\item{space}{the space to plot (matrix)}

\item{remove}{logical, the elements to remove or add (using !)}

\item{main}{the title}

\item{defaults}{a list of parameters for plot() and points()}
}
\description{
plotting the spaces changing
}
