% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdp.R
\name{plot.vdp}
\alias{plot.vdp}
\title{Make vdp plot}
\usage{
\method{plot}{vdp}(vdp, limits, pch = 19, xlab = "", ylab = "",
  disparity = NULL, ...)
}
\arguments{
\item{vdp}{A list output from \code{\link{make.vdp}}}

\item{limits}{Optional, a set of plot limits}

\item{pch}{The dots type to plot (default = 19 - full round dots)}

\item{xlab, }{ylab The x and y labels (default is none - \code{""}).}

\item{disparity}{optional, disparity values obtained from \code{\link{dispRity.vdp}} to be displayed as x labels}

\item{...}{any aditional argument to be passed to \code{\link[base]{plot}}.}
}
\description{
Make a volume, density, position combination plot
}
\examples{
## Make a Volume/density/position list
vdp_list <- make.vdp()

## Plotting the transformations
plot.vdp(vdp_list)

## Calculate disparity
vdp_disp <- dispRity.vdp(vdp_list, volume = c(sum, variances),
                                   density = c(mean, neighbours),
                                   position = c(mean, displacements))

## Plotting the results with disparity
plot.vdp(vdp_list, disparity = vdp_disp)

}
\seealso{
\code{\link{make.vdp}}, \code{\link{dispRity.vdp}}
}
\author{
Thomas Guillerme
}
