% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOconts-functions.R
\name{plotBounds}
\alias{plotBounds}
\title{Plots Stopping regions and Shows Stopping Boundaries}
\usage{
plotBounds(
  find.des.output,
  xlabel = "Stage",
  ylabel = "Test Statistic",
  title.main = NULL,
  line.df = NULL,
  ymin = -5,
  ymax = 5
)
}
\arguments{
\item{find.des.output}{The output from a call to findDes.}
}
\description{
This function plots the upper and lower stopping regions and labels the stopping boundaries,
for a design realisation found using the function findDes.
}
