% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.disparity.R
\name{prepare.disparity}
\alias{prepare.disparity}
\title{prepare.disparity}
\usage{
prepare.disparity(chain, path = "../data/")
}
\arguments{
\item{chain}{the chain name}

\item{path}{the path where the "matrices/" and "trees/" folders are}
}
\description{
Prepares data for the empirical disparity analysis (from processed data)
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
