% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.fable.R
\name{results.fable}
\alias{results.fable}
\title{Results fable}
\usage{
results.fable(all_results, test_results, col = c("orange", "blue"))
}
\arguments{
\item{all_results}{the list of all the results}

\item{test_results}{the list of the test results (named)}

\item{col}{two plotting colours}
}
\description{
Plots the results f(igure and t)able.
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
