% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.metrics.R
\name{run.reduce.spaces}
\alias{run.reduce.spaces}
\title{Reduce spaces}
\usage{
run.reduce.spaces(space, remove)
}
\arguments{
\item{space}{a space}

\item{remove}{percentage to remove}
}
\description{
Run the different space reductions
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
