% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.fable.R
\name{s.test}
\alias{s.test}
\title{Display p-value from test}
\usage{
s.test(metric.ID, type, test.results = list(s = space_test, r =
  all_dim_test))
}
\arguments{
\item{metric.ID}{The ID of the metric (i.e. row number)}

\item{type}{"s" or "r" for "space distributions" or "rarefaction"}

\item{test.results}{a list containing the results for either "s" or "r" mentioned above}
}
\description{
Displays the p-value and significance token in the fable from the tests
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
