% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift.group.simulation.R
\name{shift.group.simulation}
\alias{shift.group.simulation}
\title{Shift group}
\usage{
shift.group.simulation(space, remove, verbose = TRUE)
}
\arguments{
\item{space}{a space}

\item{remove}{percentage to remove}

\item{verbose}{whether to be verbose or not}
}
\description{
Makes shifted spaces groups for a space
}
\examples{
## Generate a space
space <- space.simulation(distributions = list(distribution = rnorm),
                         dimensions = 2, elements = 20, replicates = 1)[[1]]

## Make the shifts
shift.group.simulation(space, remove = 0.5)

}
\seealso{
space.simulation metric.simulation
}
\author{
Thomas Guillerme
}
