% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.metrics.R
\name{simulate.metrics}
\alias{simulate.metrics}
\title{Simulate metrics}
\usage{
\method{simulate}{metrics}(replicates, elements, dimensions, distributions,
  remove, metrics_list, verbose = FALSE)
}
\arguments{
\item{replicates}{number of simulations}

\item{elements}{number of elements}

\item{dimensions}{number of dimensions}

\item{distributions}{the distribution of the space}

\item{remove}{the percentage of elements to remove}

\item{metrics_list}{the list of metrics}

\item{verbose}{}
}
\description{
Run the metrics simulations
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
