% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space.simulation.R
\name{space.simulation}
\alias{space.simulation}
\title{space.simulation}
\usage{
space.simulation(distributions, dimensions, elements, replicates)
}
\arguments{
\item{distributions}{the list of arguments to be passed to \code{\link[dispRity]{space.maker}}}

\item{dimensions}{number of dimensions}

\item{elements}{number of elements}

\item{replicates}{number of simulations}
}
\description{
Wrapper function for simulating the spaces
}
\examples{
## Generate 10 normal spaces
space.simulation(distributions = list(distribution = rnorm),
                 dimensions = 2, elements = 20, replicates = 10)

## Generate 10 normal spaces with a correlation matrix and three different distributions
space.simulation(distributions = list(distribution = list(rnorm, runif, rnorm),
                                     cor.matrix = matrix(c(1,0,0,0,1,0.5,0,0.5,1), 3, 3)),
                 dimensions = 3, elements = 20, replicates = 10)

## Generate 10 random spaces with random correlations
space.simulation(distributions = list(distribution = "random", cor.matrix = "random"),
                 dimensions = 3, elements = 20, replicates = 10)

## Generate 10 random spaces with a normal scree
space.simulation(distributions = list(distribution = rnorm, scree = "normal"),
                 dimensions = 3, elements = 20, replicates = 10)

}
\seealso{
shift.space.simulation metric.simulation
}
\author{
Thomas Guillerme
}
