% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.metrics.R
\name{summarise.metrics}
\alias{summarise.metrics}
\title{Summarise metrics}
\usage{
summarise.metrics(results_list, cent.tend = median, quantiles = c(95, 50),
  scale = TRUE)
}
\arguments{
\item{results_list}{the list of results}

\item{cent.tend}{the central tendency to use}

\item{quantiles}{the quantiles to display}

\item{sale}{whether to center and scale the results}
}
\description{
Summarises the metric simulations the results for plotting and tables
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
