% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.metrics.R
\name{test.metrics}
\alias{test.metrics}
\title{Test metrics}
\usage{
test.metrics(all_results, test, scale = TRUE)
}
\arguments{
\item{test}{the test to apply}

\item{scale}{whether to scale and centre the results around the non-reduced spaces}

\item{results_list}{the list of results}
}
\description{
Applies an test to the results lists
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
