% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.simulation.R
\name{test.simulation}
\alias{test.simulation}
\title{Test metrics}
\usage{
test.simulation(results, test, scale = TRUE, factors)
}
\arguments{
\item{results}{the list of results}

\item{test}{the test to apply}

\item{scale}{whether to scale and centre the results around the non-reduced spaces}

\item{factors}{optional, a list of factors (space names) to test}
}
\description{
Applies an test to the results lists
}
\examples{
## Generate two spaces
space1 <- space.simulation(distributions = list(distribution = rnorm),
                           dimensions = 2, elements = 20, replicates = 3)
space2 <- space.simulation(distributions = list(distribution = runif),
                           dimensions = 2, elements = 20, replicates = 3)
space3 <- space.simulation(distributions = list(distribution = rlnorm),
                           dimensions = 2, elements = 20, replicates = 3)

spaces <- list("normal" = space1, "uniform" = space2, "lognorm" = space3)

## Make the shifts
groups <- lapply(spaces, lapply, shift.group.simulation, remove = 0.4)

## List of metrics
metrics <- list("sum.var" = c(sum, variances),
                "sum.ran" = c(sum, ranges))

## Apply the metrics
results <- mapply(metrics.simulation, spaces, groups,
                  MoreArgs = list(metric = metrics), SIMPLIFY = FALSE)

## Get a simple test function
test.fun <- function(data) {return(aov(glm(disparity ~ factor, data = data)))}

## Apply the test to the whole dataset
test.simulation(results, test.fun)

## Apply the test only to the normal spaces
test.simulation(results, test.fun, factors = c("normal", "lognorm"))

}
\seealso{

}
\author{
Thomas Guillerme
}
