% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdp.R
\name{vdp.check.table}
\alias{vdp.check.table}
\title{checking vdp}
\usage{
vdp.check.table(vdp_disp, vdp_space, round = 3)
}
\arguments{
\item{vdp_disp}{A list output from \code{\link{vdp.dispRity}}}

\item{vdp_space}{A list output from \code{\link{vdp.make}}}

\item{round}{The number of digits to round (default = \code{3})}
}
\description{
Checking the vdp results in a table
}
\examples{
## Make a Volume/density/position list
vdp_list <- vdp.make()

## Calculate disparity
vdp_disp <- vdp.dispRity(vdp_list, volume = c(prod, ranges),
                                   density = c(mean, neighbours),
                                   position = c(mean, displacements))

## Plotting the results with disparity
vdp.check.table(vdp_disp, vdp_list)

}
\seealso{
\code{\link{vdp.make}}, \code{\link{vdp.dispRity}}
}
\author{
Thomas Guillerme
}
