% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdp.R
\name{vdp.dispRity}
\alias{vdp.dispRity}
\title{disparity vdp}
\usage{
vdp.dispRity(vdp, volume, density, position, base.relative = TRUE)
}
\arguments{
\item{vdp}{A list output from \code{\link{vdp.make}}}

\item{volume}{a volume function as passed to \code{\link[dispRity]{dispRity}}}

\item{density}{a density function as passed to \code{\link[dispRity]{dispRity}}}

\item{position}{a position function as passed to \code{\link[dispRity]{dispRity}}}

\item{base.relative}{whether to make the results relative to the base (\code{TRUE} - default) or absolute (\code{FALSE})}
}
\description{
Measure disparity on a vdp object
}
\examples{
## Make a Volume/density/position list
vdp_list <- vdp.make()

## Plotting the transformations
vdp.plot(vdp_list)

## Calculate disparity
vdp_disp <- vdp.dispRity(vdp_list, volume = c(prod, ranges),
                                   density = c(mean, neighbours),
                                   position = c(mean, displacements))

## Plotting the results with disparity
vdp.plot(vdp_list, disparity = vdp_disp)

}
\seealso{
vdp.make, vdp.plot
}
\author{
Thomas Guillerme
}
