% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdp.R
\name{vdp.make}
\alias{vdp.make}
\title{Make vdp list}
\usage{
vdp.make(base.range = c(-0.5, 0.5), extra.points = 0)
}
\arguments{
\item{base.range}{The base range of the plots (this will be contracted and expanded; default = \code{c(-0.5, 0.5)}).}

\item{extra.points}{Wether to add any extra pair of points (default = 0).}
}
\description{
Make a volume, density, position combination list
}
\examples{
## Make a Volume/density/position list
vdp_list <- vdp.make()

## Plotting the transformations
vdp.plot(vdp_list)

## Calculate disparity
vdp_disp <- vdp.dispRity(vdp_list, volume = c(sum, variances),
                                   density = c(mean, neighbours),
                                   position = c(mean, displacements))

## Plotting the results with disparity
vdp.plot(vdp_list, disparity = vdp_disp)

}
\seealso{
vdp.plot, vdp.dispRity
}
\author{
Thomas Guillerme
}
