% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast.disparity.R
\name{fast.disparity}
\alias{fast.disparity}
\title{Fast disparity}
\usage{
fast.disparity(group, space, metric, rare.dim)
}
\arguments{
\item{group}{a logical vector for grouping}

\item{space}{a matrix}

\item{metric}{a metric dispRity style}

\item{rare.dim}{the number of dimensions to consider}
}
\description{
Fast disparity calculations (but less safe than dispRity)
}
\examples{
## A random space
space <- matrix(rnorm(25), 5, 5)

## A metric
metric <- c(sum, variances)

## A group of four observations
group <- c(TRUE, TRUE, TRUE, TRUE, FALSE)

## The disparity
fast.disparity(group, space, metric, rare.dim = 5)


}
\author{
Thomas Guillerme
}
