% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDesNewCP.R
\name{findDTL}
\alias{findDTL}
\title{Find Multi-Outcome Two-Stage Drop-the-Loser designs}
\usage{
findDTL(
  nsims = default.nsims.dtl,
  K,
  Kmax,
  m,
  vars = default.vars.dtl,
  delta0 = default.delta0.dtl,
  delta1 = default.delta1.dtl,
  delta.true = NULL,
  reuse.deltas = TRUE,
  alpha.k = default.alpha.dtl,
  alpha.combine = TRUE,
  cp.l = default.cp.l.dtl,
  cp.u = default.cp.u.dtl,
  n.min = default.nmin.dtl,
  n.max = default.nmax.dtl,
  power = default.power.dtl,
  rho.vec = default.cor.dtl,
  working.outs = NULL,
  fix.n = FALSE,
  display.lookup = TRUE
)
}
\arguments{
\item{K}{Number of outcomes}

\item{Kmax}{Maximum number of outcomes permitted in stage 2}

\item{m}{Number of outcomes required to show promise for trial success}

\item{vars}{A vector of outcome variances. If single value is entered, it is recycled with a warning.}

\item{delta0}{A vector of anticipated lower effect sizes.}

\item{delta1}{A vector of anticipated upper effect sizes.}

\item{delta.true}{Optional. A matrix of true effect sizes (with number of columns==K).}

\item{cp.l}{The lower bound for conditional power.}

\item{cp.u}{The upper bound for conditional power.}

\item{n.min}{The minimum sample size to search over.}

\item{n.max}{The maximum sample size to search over.}

\item{rho.vec}{A vector of outcome correlations. If single value is entered, it is recycled with a warning.}
}
\value{
The function returns a list of length three. The first element, input, contains the values inputted into the call.
The second element, results, gives the final and interim stopping boundaries and the operating characteristics.
The third element, lookup, is a lookup table containing the test statistics for each outcome that correspond to a range of CP values.
}
\description{
This function finds multi-outcome, two-stage drop-the-loser designs that declare trial success
when a specified number of outcomes show promise. This function uses simulation.
}
\details{
if delta.true is used, the operating characteristics of the obtained design are given, taking into account
the true effect sizes.
}
