% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.fable.R
\name{generate.fable.plot}
\alias{generate.fable.plot}
\title{Generates fable plots}
\usage{
generate.fable.plot(data, metric, what, scale = TRUE, overlap = FALSE,
  plot.param)
}
\arguments{
\item{data}{the disparity data}

\item{metric}{the metric ID}

\item{what}{the name of the chain as given pas the code snippet generating the plots}

\item{scale}{whether to scale and centre the results around the non-reduced spaces}

\item{overlap}{whether to add the Bhattacharyya Coefficients or not}

\item{plot.param}{the plot parameters}
}
\description{
Generates the plots for the fable. The R markdown compiler should name them as "<snippet_name>-<ID>.<type>"
}
\author{
Thomas Guillerme
}
