% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.utilities.R
\name{pairwise.plot}
\alias{pairwise.plot}
\title{Pairwise plots}
\usage{
pairwise.plot(results, scale = TRUE, type, factors, plot = "pairs",
  plot.param, ...)
}
\arguments{
\item{results}{the list of results}

\item{scale}{whether to scale and centre the results around the non-reduced spaces}

\item{type}{the type of results (e.g. "base" or "all")}

\item{factors}{optional, a list of factors (space names) to test}

\item{plot}{whether to display a pairplot ("pairs") or a correlation plot ("cor") or a numeric character for the metric correlations to plot.}

\item{plot.param}{plotting parameters}

\item{...}{any additional arguments to be passed to pairs() or plotcorr()}

\item{single}{logical, whether to plot single metrics (compared to all other (\code{TRUE})) or a square matrix (default - \code{FALSE})}
}
\description{
Pairwise plot of the metrics
}
