% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.utilities.R
\name{plot.metrics}
\alias{plot.metrics}
\title{Plot the results fable for the metrics analysis}
\usage{
\method{plot}{metrics}(space_results, col, remove, metrics_names, defaults,
  reduce.distribution = rnorm)
}
\arguments{
\item{space_results}{a named list of results from different simulated spaces}

\item{col}{the colours for the results (i.e. for the different spaces)}

\item{remove}{the proportion of elements removed}

\item{metrics_names}{optional, a list of metrics_names}

\item{defaults}{a list of arguments for plot.space}

\item{reduce.distribution}{the function for making the spaces for the row "names"}
}
\description{
plotting the spaces changing
}
