% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.metrics.R
\name{simulate.metrics}
\alias{simulate.metrics}
\title{Simulate metrics}
\usage{
\method{simulate}{metrics}(replicates, elements, dimensions,
  arguments = list(NULL), distributions, remove, metrics_list,
  verbose = FALSE, scree = NULL, cor.matrix = NULL, rare.dim)
}
\arguments{
\item{replicates}{number of simulations}

\item{elements}{number of elements}

\item{dimensions}{number of dimensions}

\item{distributions}{the distribution of the space}

\item{remove}{the percentage of elements to remove}

\item{metrics_list}{the list of metrics}

\item{verbose}{}

\item{scree}{variance per axis (see dispRity::space.maker)}

\item{rare.dim}{the number of dimensions to use (not to generate)}

\item{cor}{correlation between axis (see dispRity::space.maker)}
}
\description{
Run the metrics simulations
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
