% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.metrics.R
\name{summarise.metrics}
\alias{summarise.metrics}
\title{Summarise metrics}
\usage{
summarise.metrics(results_list, cent.tend = median, quantiles = c(95,
  50), scale = TRUE, relative = FALSE)
}
\arguments{
\item{results_list}{the list of results}

\item{cent.tend}{the central tendency to use}

\item{quantiles}{the quantiles to display}

\item{scale}{whether to scale the metrics (relative to the observed disparity for the full space)}

\item{relative}{whether to make the results relative to the non-changed space}
}
\description{
Summarises the metric simulations the results for plotting and tables
}
\author{
Thomas Guillerme
}
