% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.metrics.R
\name{test.metrics}
\alias{test.metrics}
\title{Test metrics}
\usage{
test.metrics()
}
\arguments{
\item{results_list}{the list of results}

\item{test}{the test to apply}

\item{relative}{whether to make the results relative to the non-changed space}
}
\description{
Applies an test to the results lists
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
